Game

Documentation

data Action

Represents different actions (including their parameters) that a cell can have on a row or column total.

Constructors

Add Int

Sub Int

data Cell

Represents a cell including whether it is enabled and its action.

Constructors

MkCell Bool Action

data Row

A row has a target number and consists of zero or more cells.

Constructors

MkRow Int [Cell]

data Grid

A grid is comprised of the target numbers for all columns and the rows.

Constructors

MkGrid [Int] [Row]

data Direction

Enumerates directions in which lists can be rotated.

Constructors

L

R

eval :: Action -> Int -> Int

eval action total applies action to the running total. For example:

>>> eval (Add 5) 3
8

>>> eval (Sub 1) 3
2

apply :: Cell -> Int -> Int

apply cell total applies the action of cell to the running total if cell is enabled. For example:

>>> apply (MkCell True (Add 5)) 3
8

>>> apply (MkCell False (Add 5)) 3
3

result :: [Cell] -> Int

result cells calculates the total produced by the actions of all enabled cells in cells starting from 0. For example:

>>> result []
0

>>> result [MkCell True (Add 5), MkCell False (Add 5), MkCell True (Sub 1)]
4

states :: Cell -> [Cell]

states cell is a function which returns a list with _exactly_ two elements that represent the two different states cell can be in. For example:

>>> states (MkCell False (Add 5))
[MkCell True (Add 5), MkCell False (Add 5)]

candidates :: [Cell] -> [[Cell]]

candidates cells is a function which, given a list of cells in a row, produces all possible combinations of states for those cells. For example:

>>> candidates [MkCell False (Add 5), MkCell False (Sub 1)]
[ [MkCell False (Add 5), MkCell False (Sub 1)]
, [MkCell False (Add 5), MkCell True (Sub 1)]
, [MkCell True (Add 5), MkCell False (Sub 1)]
, [MkCell True (Add 5), MkCell True (Sub 1)]
]

solveRow :: Row -> [Row]

solveRow row finds solutions for row. For example:

>>> solveRow (MkRow 5 [MkCell False (Add 5), MkCell False (Sub 1)])
[ MkRow 5 [MkCell True (Add 5), MkCell False (Sub 1)]]

>>> solveRow (MkRow 5 [MkCell False (Add 5), MkCell False (Add 5)])
[ MkRow 5 [MkCell True (Add 5), MkCell False (Add 5)]
, MkRow 5 [MkCell False (Add 5), MkCell True (Add 5)]
]

solve :: Grid -> [Grid]

solve grid finds all solutions for grid. For example:

>>> let row0 = MkRow 3 [MkCell False (Add 3), MkCell False (Add 5)]
>>> let row1 = MkRow 4 [MkCell False (Add 2), MkCell False (Add 2)]
>>> solve (MkGrid [5, 2] [row0, row1])
[ MkGrid [5,2] [ MkRow 3 [MkCell True (Add 3), MkCell False (Add 5)]
               , MkRow 4 [MkCell True (Add 2), MkCell True (Add 2)]
               ]
]

rotate :: Direction -> [a] -> [a]

rotate direction list rotates the items in list to the left or right depending on the value of direction. For example:

>>> rotate L [1,2,3]
[2,3,1]

>>> rotate R [1,2,3]
[3,1,2]

>>> rotate L []
[]

>>> rotate R [1]
[1]

rotations :: Grid -> [Grid]

rotations grid returns a list of grids containing all possible ways to rotate grid. This means the resulting list should normally have rows + columns many elements. For example:

>>> let row0 = MkRow 3 [MkCell False (Add 3), MkCell False (Add 5)]
>>> let row1 = MkRow 4 [MkCell False (Add 2), MkCell False (Add 2)]
>>> rotations (MkGrid [5, 2] [row0, row1])
[ MkGrid [5,2] [ MkRow 3 [MkCell False (Add 5), MkCell False (Add 3)]
               , MkRow 4 [MkCell False (Add 2), MkCell False (Add 2)]
               ]
, MkGrid [5,2] [ MkRow 3 [MkCell False (Add 3), MkCell False (Add 5)]
               , MkRow 4 [MkCell False (Add 2), MkCell False (Add 2)]
               ]
, MkGrid [5,2] [ MkRow 3 [MkCell False (Add 2), MkCell False (Add 5)]
               , MkRow 4 [MkCell False (Add 3), MkCell False (Add 2)]
               ]
, MkGrid [5,2] [ MkRow 3 [MkCell False (Add 3), MkCell False (Add 2)]
               , MkRow 4 [MkCell False (Add 2), MkCell False (Add 5)]
               ]
]

steps :: Grid -> [Grid]

steps grid finds the sequence of rotations that lead to a solution for grid in the fewest number of rotations. The resulting list includes the solution as the last element. You may assume that this function will never be called on a grid for which there are solutions returned by solve. The states of intermediate grids in the resulting list are irrelevant - only the ones of the final grid need to be set correctly.

>>> let row0 = MkRow 3 [MkCell False (Add 2), MkCell False (Add 3)]
>>> let row1 = MkRow 4 [MkCell False (Add 5), MkCell False (Add 2)]
>>> steps (MkGrid [5, 2] [row0, row1])
[ MkGrid [5, 2] [ MkRow 3 [ MkCell False (Add 5), MkCell False (Add 3)]
                , MkRow 4 [ MkCell False (Add 2), MkCell False (Add 2)]
                ]
, MkGrid [5, 2] [ MkRow 3 [ MkCell True (Add 3), MkCell False (Add 5)]
                , MkRow 4 [ MkCell True (Add 2), MkCell True (Add 2)]
                ]
]