Level

Description

This module contains types to represent levels and campaigns along with functions to load/save levels and campaigns.

Documentation

data Level

Represents a level consisting of a par number of rotations and a grid.

Constructors

MkLevel

Fields

levelPar :: Int

The par number of rotations required to solve this level.

levelGrid :: Grid

The grid that makes up the level.

loadLevel :: FilePath -> IO (Either String Level)

loadLevel filepath tries to load a level from filepath.

saveLevel :: FilePath -> Level -> IO ()

saveLevel filepath level saves level to filepath.

data Campaign

Represents a campaign (a collection of levels).

Constructors

MkCampaign

Fields

campaignTitle :: String

The title of the campaign.

campaignDesc :: String

The description of the campaign.

campaignPath :: FilePath

The path to the levels folder for this campaign.

listCampaigns :: IO (Maybe [Campaign])

listCampaigns is a computation which loads the list of campaigns.

loadLevels :: FilePath -> IO (Either String [Level])

loadLevels directory attempts to load all levels from directory.

Orphan instances

ToJSON Grid

Instance details

Methods

toJSON :: Grid -> Value

toEncoding :: Grid -> Encoding

toJSONList :: [Grid] -> Value

toEncodingList :: [Grid] -> Encoding

ToJSON Row

Instance details

Methods

toJSON :: Row -> Value

toEncoding :: Row -> Encoding

toJSONList :: [Row] -> Value

toEncodingList :: [Row] -> Encoding

ToJSON Cell

Instance details

Methods

toJSON :: Cell -> Value

toEncoding :: Cell -> Encoding

toJSONList :: [Cell] -> Value

toEncodingList :: [Cell] -> Encoding

ToJSON Action

Instance details

Methods

toJSON :: Action -> Value

toEncoding :: Action -> Encoding

toJSONList :: [Action] -> Value

toEncodingList :: [Action] -> Encoding

FromJSON Grid

Instance details

Methods

parseJSON :: Value -> Parser Grid

parseJSONList :: Value -> Parser [Grid]

FromJSON Row

Instance details

Methods

parseJSON :: Value -> Parser Row

parseJSONList :: Value -> Parser [Row]

FromJSON Cell

Instance details

Methods

parseJSON :: Value -> Parser Cell

parseJSONList :: Value -> Parser [Cell]

FromJSON Action

Instance details

Methods

parseJSON :: Value -> Parser Action

parseJSONList :: Value -> Parser [Action]